requirements.txt
asyncstdlib
black
datetime
pylint
pyyaml
telethon
yamllint
